+++ fabric/src/main/java/de/cech12/usefulhats/client/UsefulHatsFabricClientMod.java
@@ -1,6 +1,8 @@
 package de.cech12.usefulhats.client;
 
 import de.cech12.usefulhats.Constants;
+import de.cech12.usefulhats.client.compat.AccessoriesClientCompat;
+import de.cech12.usefulhats.compat.AccessoriesCompat;
 import de.cech12.usefulhats.item.AbstractHatItem;
 import de.cech12.usefulhats.platform.Services;
 import net.fabricmc.api.ClientModInitializer;
@@ -34,6 +36,9 @@ public class UsefulHatsFabricClientMod implements ClientModInitializer {
             TrinketsClientCompat.register();
         }
          */
+        if (Services.PLATFORM.isModLoaded(AccessoriesCompat.MOD_ID)) {
+            AccessoriesClientCompat.register();
+        }
     }
 
     public static void resetUsefulHatModel() {